refactor RPS and Connect Four
